Clean up sparse_tensor duplication in LAPIS
sparse_tensor namespace stuff should only live in LLVM upstream, not
LAPIS. Now all kokkos and part_tensor passes/pipelines live in their
respective namespaces.
