+++ mlir/include/mlir/Dialect/Kokkos/Transforms/Passes.h
@@ -38,6 +38,11 @@ std::unique_ptr<Pass> createKokkosMemorySpaceAssignmentPass();
 void populateKokkosDualViewManagementPatterns(RewritePatternSet &patterns);
 std::unique_ptr<Pass> createKokkosDualViewManagementPass();
 
+// Old Kokkos codegen for kernel outlining and host-device copying
+// (does not use the ops in Kokkos dialect)
+void populateSparseKokkosCodegenPatterns(RewritePatternSet &patterns);
+std::unique_ptr<Pass> createSparseKokkosCodegenPass();
+
 //===----------------------------------------------------------------------===//
 // Registration.
 //===----------------------------------------------------------------------===//
