+++ mlir/include/mlir/Dialect/Kokkos/Transforms/Passes.td
@@ -61,4 +61,22 @@ def KokkosDualViewManagement : Pass<"kokkos-dualview-management"> {
   ];
 }
 
+// Old Kokkos GPU codegen (not using Kokkos dialect ops)
+def SparseKokkosCodegen : Pass<"sparse-kokkos-codegen", "ModuleOp"> {
+  let summary = "Generates Kokkos code during sparsification";
+  let description = [{
+    Enables sparse compiler to generate Kokkos code, which can use non-host backends.
+  }]; 
+  let constructor = "mlir::createSparseKokkosCodegenPass()";
+  let dependentDialects = [ 
+    "arith::ArithDialect",
+    "bufferization::BufferizationDialect",
+    "emitc::EmitCDialect",
+    "linalg::LinalgDialect",
+    "memref::MemRefDialect",
+    "scf::SCFDialect",
+    "sparse_tensor::SparseTensorDialect",
+  ];  
+}
+
 #endif // MLIR_DIALECT_KOKKOS_TRANSFORMS_PASSES
