--- mlir/include/mlir/Dialect/SparseTensor/CMakeLists.txt
@@ -1,1 +1,0 @@
-add_subdirectory(Transforms)
