+++ mlir/lib/CAPI/Registration.cpp
@@ -1,10 +1,10 @@
 #include "mlir/InitAllKokkosPasses.h"
 #include "mlir-c/kmDialects.h"
 #include "mlir/CAPI/Registration.h"
-#include "mlir/Dialect/SparseTensor/IR/SparseTensor.h"
+#include "mlir/Dialect/Kokkos/IR/KokkosDialect.h"
 
 void lapisRegisterAllPasses() {
   mlir::registerAllKokkosPasses();
 }
 
-MLIR_DEFINE_CAPI_DIALECT_REGISTRATION(Kokkos, kokkos, mlir::sparse_tensor::SparseTensorDialect)
+MLIR_DEFINE_CAPI_DIALECT_REGISTRATION(Kokkos, kokkos, mlir::kokkos::KokkosDialect)
