--- mlir/lib/Dialect/SparseTensor/KokkosPipelines/CMakeLists.txt
@@ -1,28 +1,0 @@
-add_mlir_dialect_library(MLIRSparseTensorKokkosPipelines
-SparseTensorKokkosPipelines.cpp
-
-  ADDITIONAL_HEADER_DIRS
-  ${MLIR_MAIN_INCLUDE_DIR}/mlir/Dialect/SparseTensor
-
-  LINK_LIBS PUBLIC
-  MLIRArithTransforms
-  MLIRAffineToStandard
-  MLIRBufferizationTransforms
-  MLIRComplexToLLVM
-  MLIRComplexToLibm
-  MLIRComplexToStandard
-  MLIRFuncTransforms
-  MLIRLinalgTransforms
-  MLIRMathToLibm
-  MLIRMathToLLVM
-  MLIRMemRefToLLVM
-  LAPISPass
-  MLIRReconcileUnrealizedCasts
-  MLIRSCFToControlFlow
-  MLIRSparseTensorDialect
-  MLIRSparseTensorTransforms
-  MLIRTensorTransforms
-  MLIRVectorToLLVM
-  MLIRVectorTransforms
-  Kokkos_MLIRSparseTensorTransforms
-)
