--- mlir/lib/Dialect/SparseTensor/KokkosPipelines/SparseTensorKokkosPipelines.cpp
@@ -1,71 +1,0 @@
-//===- SparseTensorKokkosPipelines.cpp - Pipelines for sparse tensor code -===//
-//
-// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
-// See https://llvm.org/LICENSE.txt for license information.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-//
-//===----------------------------------------------------------------------===//
-
-#include "mlir/Dialect/SparseTensor/Pipelines/kmPasses.h"
-
-#include "mlir/Conversion/Passes.h"
-#include "mlir/Dialect/Arith/Transforms/Passes.h"
-#include "mlir/Dialect/Bufferization/Transforms/Bufferize.h"
-#include "mlir/Dialect/Bufferization/Transforms/OneShotAnalysis.h"
-#include "mlir/Dialect/Bufferization/Transforms/Passes.h"
-#include "mlir/Dialect/Func/IR/FuncOps.h"
-#include "mlir/Dialect/Linalg/Passes.h"
-#include "mlir/Dialect/MemRef/Transforms/Passes.h"
-#ifdef ENABLE_PART_TENSOR
-#include "mlir/Dialect/PartTensor/Transforms/Passes.h"
-#endif
-#include "mlir/Dialect/SparseTensor/IR/SparseTensor.h"
-#include "mlir/Dialect/SparseTensor/Transforms/kmPasses.h"
-#include "mlir/Pass/kmPassManager.h"
-#include "mlir/Transforms/Passes.h"
-
-using namespace mlir;
-using namespace mlir::sparse_tensor;
-
-//===----------------------------------------------------------------------===//
-// Pipeline implementation.
-//===----------------------------------------------------------------------===//
-
-void mlir::sparse_tensor::buildSparseKokkosCompiler(
-    OpPassManager &pm, const SparseCompilerOptions &options) {
-#ifdef ENABLE_PART_TENSOR
-  pm.addPass(createPartTensorConversionPass());
-#endif
-  pm.addNestedPass<func::FuncOp>(createLinalgGeneralizationPass());
-  pm.addPass(createSparsificationAndBufferizationPass(
-      getBufferizationOptionsForSparsification(
-          options.testBufferizationAnalysisOnly),
-      options.sparsificationOptions(), options.sparseTensorConversionOptions(),
-      options.createSparseDeallocs, options.enableRuntimeLibrary,
-      options.enableBufferInitialization, options.vectorLength,
-      /*enableVLAVectorization=*/options.armSVE,
-      /*enableSIMDIndex32=*/options.force32BitVectorIndices));
-  pm.addNestedPass<func::FuncOp>(createCanonicalizerPass());
-  pm.addNestedPass<func::FuncOp>(
-      mlir::bufferization::createFinalizingBufferizePass());
-  pm.addPass(createSparseKokkosCodegenPass());
-  pm.addNestedPass<func::FuncOp>(createConvertLinalgToLoopsPass());
-  pm.addNestedPass<func::FuncOp>(createConvertVectorToSCFPass());
-  pm.addNestedPass<func::FuncOp>(memref::createExpandReallocPass());
-  pm.addPass(memref::createExpandStridedMetadataPass());
-  pm.addPass(createLowerAffinePass());
-  pm.addPass(createReconcileUnrealizedCastsPass());
-}
-
-//===----------------------------------------------------------------------===//
-// Pipeline registration.
-//===----------------------------------------------------------------------===//
-
-void mlir::sparse_tensor::registerSparseTensorKokkosPipelines() {
-  PassPipelineRegistration<SparseCompilerOptions>(
-      "sparse-compiler-kokkos",
-      "The standard pipeline for taking sparsity-agnostic IR using the"
-      " sparse-tensor type, and lowering it to dialects compatible with"
-      " the Kokkos emitter.",
-      buildSparseKokkosCompiler);
-}
