+++ mlir/unittests/Dialect/CMakeLists.txt
@@ -9,7 +9,6 @@ target_link_libraries(MLIRDialectTests
 add_subdirectory(LLVMIR)
 add_subdirectory(MemRef)
 add_subdirectory(PartTensor)
-add_subdirectory(SparseTensor)
 add_subdirectory(SPIRV)
 add_subdirectory(Transform)
 add_subdirectory(Utils)
